@@ -12,12 +12,12 @@
     "outdated": "ncu -i"
   },
   "devDependencies": {
-    "@apollo/client": "^3.10.4",
-    "@eslint/js": "^9.4.0",
+    "@apollo/client": "^3.10.5",
+    "@eslint/js": "^9.5.0",
     "@popperjs/core": "^2.11.8",
     "@rails/actioncable": "^7.1.3",
-    "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/kit": "^2.5.10",
+    "@sveltejs/adapter-static": "^3.0.2",
+    "@sveltejs/kit": "^2.5.16",
     "@sveltejs/vite-plugin-svelte": "^3.1.1",
     "@types/bootstrap": "^5.2.10",
     "@types/eslint": "^8.56.10",
@@ -25,23 +25,23 @@
     "@types/qrcode": "^1.5.5",
     "@types/rails__actioncable": "^6.1.10",
     "bootstrap": "^5.3.3",
-    "eslint": "^9.4.0",
+    "eslint": "^9.5.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.39.0",
-    "graphql": "^16.8.1",
+    "eslint-plugin-svelte": "^2.39.5",
+    "graphql": "^16.8.2",
     "graphql-ruby-client": "^1.13.3",
     "npm-check-updates": "^16.14.20",
-    "prettier": "^3.3.1",
-    "prettier-plugin-svelte": "^3.2.3",
+    "prettier": "^3.3.2",
+    "prettier-plugin-svelte": "^3.2.4",
     "qrcode": "^1.5.3",
-    "sass": "^1.77.4",
-    "svelte": "^4.2.17",
+    "sass": "^1.77.5",
+    "svelte": "^5.0.0-next.155",
     "svelte-check": "^3.8.0",
     "svelte-i18n": "^4.0.0",
-    "svelte-preprocess": "^5.1.4",
+    "svelte-preprocess": "^6.0.1",
     "tslib": "^2.6.3",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.12.0",
-    "vite": "^5.2.12"
+    "typescript-eslint": "^7.13.0",
+    "vite": "^5.3.1"
   }
 }
