@@ -1,19 +1,22 @@
 <script lang="ts">
   import Modal from 'bootstrap/js/dist/modal'
-  import { createEventDispatcher } from 'svelte'
-  const dispatch = createEventDispatcher()
+
   import { _ } from '$lib/helpers/i18n'
+
+  interface Props {
+    onconfirm: () => void
+    title?: string
+  }
+
+  let { onconfirm, title = $_('components.confirm.title') }: Props = $props()
 
   let dialog: HTMLDivElement
 
   function confirm() {
     var modal = Modal.getInstance(dialog)
     modal?.hide()
-
-    dispatch('confirm')
+    onconfirm()
   }
-
-  export let title = $_('components.confirm.title')
 </script>
 
 <div class="modal fade" id="confirmDialog" bind:this={dialog}>
@@ -21,13 +24,14 @@
     <div class="modal-content">
       <div class="modal-header">
         <h5 class="modal-title">{title}</h5>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" />
+        <button type="button" class="btn-close" data-bs-dismiss="modal"
+        ></button>
       </div>
       <div class="modal-footer">
         <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">
           {$_('components.confirm.cancel')}
         </button>
-        <button type="button" class="btn btn-danger" on:click={confirm}>
+        <button type="button" class="btn btn-danger" onclick={confirm}>
           {$_('components.confirm.proceed')}
         </button>
       </div>
