@@ -1,16 +1,19 @@
 <script lang="ts">
-  import { onMount } from 'svelte'
   import QRCode from 'qrcode'
 
-  export let content: string
+  interface Props {
+    content: string
+  }
+
+  let { content }: Props = $props()
 
   let canvas: HTMLElement
 
-  onMount(() => {
+  $effect(() => {
     QRCode.toCanvas(canvas, content, {
       width: 300
     })
   })
 </script>
 
-<canvas id="qrcode" bind:this={canvas} />
+<canvas id="qrcode" bind:this={canvas}></canvas>
