@@ -11,28 +11,30 @@
 
   import type { Post } from '$lib/types/Post'
 
-  let uuid = $page.params.uuid
+  let uuid = $derived($page.params.uuid)
 
-  let post: Post
+  let post: Post | undefined = $state()
 
   const fields = `uuid title body user { uuid name }`
 
-  findPost({ uuid }, fields)
-    .then(res => {
-      post = res.data.findPost
-    })
-    .catch(error => {
-      errors.set(error.graphQLErrors)
-    })
+  $effect(() => {
+    findPost({ uuid }, fields)
+      .then(res => {
+        post = res.data.findPost
+      })
+      .catch(error => {
+        errors.set(error.graphQLErrors)
+      })
 
-  watchPost({ uuid }, fields).subscribe(
-    res => {
-      if (res.data.postUpdated) post = res.data.postUpdated
-    },
-    err => {
-      errors.set(err.graphQLErrors)
-    }
-  )
+    watchPost({ uuid }, fields).subscribe(
+      res => {
+        if (res.data.postUpdated) post = res.data.postUpdated
+      },
+      err => {
+        errors.set(err.graphQLErrors)
+      }
+    )
+  })
 
   function destroy() {
     destroyPost({ uuid }, `post { uuid }`)
@@ -45,33 +47,35 @@
   }
 </script>
 
-<Loader loaded={!!post}>
-  <h1>{post.title}</h1>
-  <p>
-    <small
-      >By <a href="/users/{post.user.uuid}">
-        {post.user.name}
-      </a></small
-    >
-  </p>
-  <p>{post.body}</p>
-
-  {#if $session.user && $session.user.uuid == post.user.uuid}
+<Loader>
+  {#if post}
+    <h1>{post.title}</h1>
     <p>
-      <a href="/posts/{post.uuid}/edit" class="btn btn-outline-primary">
-        {$_('common.edit')}
-      </a>
-
-      <button
-        type="button"
-        class="btn btn-outline-danger"
-        data-bs-toggle="modal"
-        data-bs-target="#confirmDialog"
+      <small
+        >By <a href="/users/{post.user.uuid}">
+          {post.user.name}
+        </a></small
       >
-        {$_('common.delete')}
-      </button>
     </p>
+    <p>{post.body}</p>
+
+    {#if $session.user && $session.user.uuid == post.user.uuid}
+      <p>
+        <a href="/posts/{post.uuid}/edit" class="btn btn-outline-primary">
+          {$_('common.edit')}
+        </a>
+
+        <button
+          type="button"
+          class="btn btn-outline-danger"
+          data-bs-toggle="modal"
+          data-bs-target="#confirmDialog"
+        >
+          {$_('common.delete')}
+        </button>
+      </p>
+    {/if}
   {/if}
 </Loader>
 
-<Confirm on:confirm={destroy} />
+<Confirm onconfirm={destroy} />
