@@ -1,4 +1,4 @@
-export type PageInfo = {
+export interface PageInfo {
   startCursor: string
   endCursor: string
   hasPreviousPage: string
