@@ -6,14 +6,16 @@
   import { errors } from '$lib/helpers/stores'
   import { setLocale } from '$lib/helpers/i18n'
 
-  logoutUser({}, `user { uuid }`)
-    .then(() => {
-      clearSession()
-      setLocale.set('en')
+  $effect(() => {
+    logoutUser({}, `user { uuid }`)
+      .then(() => {
+        clearSession()
+        setLocale.set('en')
 
-      goto('/')
-    })
-    .catch(error => {
-      errors.set(error.graphQLErrors)
-    })
+        goto('/')
+      })
+      .catch(error => {
+        errors.set(error.graphQLErrors)
+      })
+  })
 </script>
