@@ -10,20 +10,24 @@
 
   import type { User } from '$lib/types/User'
 
-  let uuid = $page.params.uuid
+  let uuid = $derived($page.params.uuid)
 
-  let user: User
-  let code: string
+  let user: User | undefined = $state()
+  let code: string = $state('')
 
-  findUser({ uuid }, `uuid otpEnabled otpProvisioningUrl`)
-    .then(res => {
-      user = res.data.findUser
-    })
-    .catch(error => {
-      errors.set(error.graphQLErrors)
-    })
+  $effect(() => {
+    findUser({ uuid }, `uuid otpEnabled otpProvisioningUrl`)
+      .then(res => {
+        user = res.data.findUser
+      })
+      .catch(error => {
+        errors.set(error.graphQLErrors)
+      })
+  })
 
-  function submitEnable() {
+  function submitEnable(event: SubmitEvent) {
+    event.preventDefault()
+
     updateUser(
       {
         uuid,
@@ -43,7 +47,9 @@
       })
   }
 
-  function submitDisable() {
+  function submitDisable(event: SubmitEvent) {
+    event.preventDefault()
+
     updateUser(
       {
         uuid,
@@ -68,52 +74,54 @@
   {$_('pages.users.mfa.heading')}
 </h1>
 
-<Loader loaded={!!user}>
-  {#if user.otpEnabled}
-    <form on:submit|preventDefault={submitDisable}>
-      <div class="mb-3">
-        <label for="code" class="form-label">
-          {$_('pages.users.mfa.code')}
-        </label>
-        <input
-          type="text"
-          class="form-control"
-          id="name"
-          bind:value={code}
-          required
-        />
-      </div>
+<Loader>
+  {#if user}
+    {#if user.otpEnabled}
+      <form onsubmit={submitDisable}>
+        <div class="mb-3">
+          <label for="code" class="form-label">
+            {$_('pages.users.mfa.code')}
+          </label>
+          <input
+            type="text"
+            class="form-control"
+            id="name"
+            bind:value={code}
+            required
+          />
+        </div>
 
-      <div class="mb-3">
-        <button type="submit" class="btn btn-primary">
-          {$_('pages.users.mfa.disable')}
-        </button>
-      </div>
-    </form>
-  {:else}
-    <form on:submit|preventDefault={submitEnable}>
-      <div class="md-3">
-        <QRCode content={user.otpProvisioningUrl} />
-      </div>
+        <div class="mb-3">
+          <button type="submit" class="btn btn-primary">
+            {$_('pages.users.mfa.disable')}
+          </button>
+        </div>
+      </form>
+    {:else}
+      <form onsubmit={submitEnable}>
+        <div class="md-3">
+          <QRCode content={user.otpProvisioningUrl} />
+        </div>
 
-      <div class="mb-3">
-        <label for="name" class="form-label">
-          {$_('pages.users.mfa.code')}
-        </label>
-        <input
-          type="text"
-          class="form-control"
-          id="name"
-          bind:value={code}
-          required
-        />
-      </div>
+        <div class="mb-3">
+          <label for="name" class="form-label">
+            {$_('pages.users.mfa.code')}
+          </label>
+          <input
+            type="text"
+            class="form-control"
+            id="name"
+            bind:value={code}
+            required
+          />
+        </div>
 
-      <div class="mb-3">
-        <button type="submit" class="btn btn-primary">
-          {$_('pages.users.mfa.enable')}
-        </button>
-      </div>
-    </form>
+        <div class="mb-3">
+          <button type="submit" class="btn btn-primary">
+            {$_('pages.users.mfa.enable')}
+          </button>
+        </div>
+      </form>
+    {/if}
   {/if}
 </Loader>
